almost done with the tree logic
pretty much finished the traversal logic just need to fix some bugs here and there and add the lvls to the traversal. Character has been added to the tree min game but still need the virtual character added and the logic for getting kicked back to the lobby has been implemented. when testing out the lives wait for 5 seconds when inputting a new sequence for there is a bug at the moment that i will work on later